webApp/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
openai
faker
groq
python-dotenv
langchain
sentence-transformers
jinja2